placement: fix node best cost estimation
place_irq_in_node() attempts to find the node with the lowest load by
comparing them with best_cost.  Here we were initializing best_cost with
INT_MAX but our loads are computed as unsigned long long and thus may
all be greater than best_cost.  This initializes best cost correctly.

The symptom was that if nodes were under heavy load and an irq was
selected for rebalancing it would perpetually remain on the
rebalance_irq_list.  This is the same issue that was partially fixed in
commit b89f13cff17bd75b811aec5b3f97768f47f93472 "placement: Fix best
cost estimation"

+++ placement.c
@@ -153,7 +153,7 @@ static void place_irq_in_node(struct irq_info *info, void *data __attribute__((u
 		return;
 	}
 
-	place.best_cost = INT_MAX;
+	place.best_cost = ULLONG_MAX;
 	place.best = NULL;
 	place.least_irqs = NULL;
 	place.info = info;
